+++ lib/bloom_net_center_client/client.rb
@@ -2,7 +2,9 @@ module BloomNetCenterClient
   class Client
 
     include Virtus.model
-    attribute :host, String
+    BloomNetCenterClient::GLOBAL_OPTS.each do |attr|
+      attribute attr, String
+    end
 
     include APIClientBase::Client.module(default_opts: :default_opts)
     api_action :create_txn
@@ -11,7 +13,9 @@ module BloomNetCenterClient
     private
 
     def default_opts
-      { host: host }
+      BloomNetCenterClient::GLOBAL_OPTS.each_with_object({}) do |attr, hash|
+        hash[attr] = send(attr)
+      end
     end
 
   end