+++ spec/lib/bloom_net_center_client/client_spec.rb
@@ -1,7 +1,14 @@
 require "spec_helper"
 
 module BloomNetCenterClient
-  RSpec.describe Client do
+  RSpec.describe Client, type: %i[virtus] do
+
+    describe "attributes" do
+      subject { described_class }
+      BloomNetCenterClient::GLOBAL_OPTS.each do |attr|
+        it { is_expected.to have_attribute(attr, String) }
+      end
+    end
 
   end
 end