+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Changed
+- Work with new version of BloomNet Center
+
 ## [0.2.0] - 2017-01-20
 ### Changed
 - Update api_client_base to `~> 0.2.0`